requirements.txt
pytest
selenium
webdriver-manager
python-dotenv
